refactor, cleanup: domain mapper in resolver, can get mappings from gql, only the custom domain's sub can be selected during post, middleware cleanup, test domain logger
