+++ .env.development
@@ -194,3 +194,6 @@ NEXT_TELEMETRY_DISABLED=1
 
 # DNS resolver for custom domain verification
 DNS_RESOLVER=1.1.1.1
+
+# domain debug
+DOMAIN_DEBUG=true
